+++ requests/__init__.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from . import packages
-from .core import *
+import packages
+from core import *
+
+from core import __version__